+++ yukka/src/main/java/com/example/yukka/model/social/service/KomentarzService.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,6 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.example.yukka.common.PageResponse;
 import com.example.yukka.file.FileStoreService;
+import com.example.yukka.file.FileUtils;
 import com.example.yukka.handler.BannedUzytkownikException;
 import com.example.yukka.handler.EntityNotFoundException;
 import com.example.yukka.model.social.komentarz.Komentarz;
@@ -50,6 +52,7 @@ public class KomentarzService {
     private final RozmowaPrywatnaRepository rozmowaPrywatnaRepository;
     private final KomentarzRepository komentarzRepository;
     private final FileStoreService fileStoreService;
+    private final FileUtils fileUtils;
     private final PowiadomienieService powiadomienieService;
 
    // PostMapper postMapper;
@@ -369,6 +372,32 @@ public class KomentarzService {
         return response;
     }
 
+    public Komentarz addOdpowiedzToKomentarz(@Valid KomentarzRequest request, MultipartFile file, Uzytkownik connectedUser) throws FileUploadException {
+        
+        Uzytkownik uzyt = connectedUser;
+        Optional<Komentarz> newestKomentarz = komentarzRepository.findNewestKomentarzOfUzytkownik(uzyt.getEmail());
+        checkTimeSinceLastKomentarz(newestKomentarz);
+
+        Post post = postRepository.findPostByKomentarzOdpowiedzId(request.getTargetId()).orElseThrow(() -> new EntityNotFoundException("Nie znaleziono posta dla odpowiedzi o podanym ID: " + request.getTargetId()));
+        Komentarz komentarzDoOdpowiedzi = komentarzRepository.findKomentarzByKomentarzId(request.getTargetId()).orElseThrow(() -> new EntityNotFoundException("Nie znaleziono komentarza o podanym ID: " + request.getTargetId()));
+
+        Komentarz kom = createKomentarz(request);
+        saveKomentarzFile(file, kom, uzyt);
+        
+        Komentarz response = komentarzRepository.addOdpowiedzToKomentarzInPost(uzyt.getEmail(), kom, request.getTargetId());
+
+        if(!komentarzDoOdpowiedzi.getUzytkownik().getUzytId().equals(uzyt.getUzytId())) {
+            PowiadomienieResponse powiadomienie = PowiadomienieResponse.builder()
+            .typ(TypPowiadomienia.KOMENTARZ_POST.name()).tytul(post.getTytul())
+            .odnosnik(post.getPostId()).uzytkownikNazwa(uzyt.getNazwa())
+            .avatar(uzyt.getAvatar())
+            .build();
+            powiadomienieService.addPowiadomienie(powiadomienie, komentarzDoOdpowiedzi.getUzytkownik());
+        }
+
+        return response;
+    }
+
     public Komentarz updateKomentarz(String komentarzId, @Valid KomentarzRequest request, Authentication connectedUser) {
         Uzytkownik uzyt = ((Uzytkownik) connectedUser.getPrincipal());
         Komentarz kom = komentarzRepository.findKomentarzByKomentarzId(komentarzId)
@@ -381,19 +410,32 @@ public class KomentarzService {
     public void deleteKomentarz(String komentarzId, Authentication connectedUser) {
         Uzytkownik uzyt = ((Uzytkownik) connectedUser.getPrincipal());        
         Komentarz komentarz = komentarzRepository.findKomentarzByKomentarzId(komentarzId)
-                .filter(k -> uzyt.hasAuthenticationRights(k.getUzytkownik(), connectedUser))
-                .orElseThrow();
+            .orElseThrow();
+        
+        if (uzyt.hasAuthenticationRights(komentarz.getUzytkownik(), connectedUser)) {
+            throw new AccessDeniedException("Nie masz uprawnień do usunięcia komentarza");
+        }
+
+        fileUtils.deleteObraz(komentarz.getObraz());
         komentarzRepository.removeKomentarz(komentarz.getKomentarzId());
     }
 
     public void deleteKomentarzFromPost(String postId, String komentarzId, Authentication connectedUser) {
         Uzytkownik uzyt = ((Uzytkownik) connectedUser.getPrincipal());
         Post post = postRepository.findPostByPostId(postId).orElseThrow(() -> new EntityNotFoundException("Nie znaleziono posta o podanym ID: " + postId));
-        Komentarz kom = komentarzRepository.findKomentarzByKomentarzId(komentarzId)
-                .filter(k -> uzyt.hasAuthenticationRights(k.getUzytkownik(), connectedUser))
+        Komentarz kom = komentarzRepository.findKomentarzWithOdpowiedziByKomentarzId(komentarzId)
                 .orElseThrow(() -> new EntityNotFoundException("Nie znaleziono komentarza o podanym ID: " + komentarzId));
+        
+        if (uzyt.hasAuthenticationRights(kom.getUzytkownik(), connectedUser)) {
+            throw new AccessDeniedException("Nie masz uprawnień do usunięcia komentarza");
+        }
 
         List<Uzytkownik> uzytkownicyInPost = uzytkownikRepository.getConnectedUzytkownicyFromPostButBetter(post.getPostId());
+
+        fileUtils.deleteObraz(kom.getObraz());
+        for (Komentarz odp : kom.getOdpowiedzi()) {
+            fileUtils.deleteObraz(odp.getObraz());
+        }
         komentarzRepository.removeKomentarz(kom.getKomentarzId());
         
         komentarzRepository.updateUzytkownikKomentarzeOcenyCount(uzytkownicyInPost);
@@ -403,11 +445,16 @@ public class KomentarzService {
     public void deleteKomentarzFromPost(String postId, String komentarzId, Uzytkownik connectedUser) {
         Uzytkownik uzyt = connectedUser;
         Post post = postRepository.findPostByPostId(postId).orElseThrow(() -> new EntityNotFoundException("Nie znaleziono posta o podanym ID: " + postId));
-        Komentarz kom = komentarzRepository.findKomentarzByKomentarzId(komentarzId).orElseThrow(() -> new EntityNotFoundException("Nie znaleziono komentarza o podanym ID: " + komentarzId));
+        Komentarz kom = komentarzRepository.findKomentarzWithOdpowiedziByKomentarzId(komentarzId).orElseThrow(() -> new EntityNotFoundException("Nie znaleziono komentarza o podanym ID: " + komentarzId));
 
         System.out.println("Pobieranie użytowników z posta");
       //  List<Uzytkownik> uzytkownicyInPost = postRepository.getConnectedUzytkownicyFromPostButBetter(post.getPostId());
         List<Uzytkownik> uzytkownicyInPost = uzytkownikRepository.getConnectedUzytkownicyFromPostButBetter(post.getPostId());
+
+        fileUtils.deleteObraz(kom.getObraz());
+        for (Komentarz odp : kom.getOdpowiedzi()) {
+            fileUtils.deleteObraz(odp.getObraz());
+        }
         komentarzRepository.removeKomentarz(kom.getKomentarzId());
 
         System.out.println("Aktualizacja użytowników");
@@ -422,6 +469,16 @@ public class KomentarzService {
 
         System.out.println("Aktualizacja posta");
         komentarzRepository.updateKomentarzeCountInPostButWithFunniNewRelation(post.getPostId());
+    }
+
+    // Seedowane bo trzeba usuwać obrazy
+    public void seedRemoveKomentarzeObrazy() {
+        List<Komentarz> komentarze = komentarzRepository.findAll();
+        System.out.println("Usuwanie obrazów komentarzy");
+        for (Komentarz kom : komentarze) {
+            System.out.println("Usuwanie obrazu komentarza: " + kom.getKomentarzId());
+            fileUtils.deleteObraz(kom.getObraz());
+        }
     }
 
     // Pomocnicze
@@ -455,7 +512,6 @@ public class KomentarzService {
     }
 
     private void checkTimeSinceLastKomentarz(Optional<Komentarz> newestKomentarz) {
-        
         if (newestKomentarz.isPresent()) {
             LocalDateTime lastKomentarzTime = newestKomentarz.get().getDataUtworzenia();
             LocalDateTime now = LocalDateTime.now();
